Refactor error handling into `model/utils/errors.ts`
Moved error-related functions and classes from `prompt` to the
`model/utils/errors.ts` file for better organization and reuse. Updated
imports to reflect this change.
